requirements.txt
blinker==1.9.0
click==8.3.0
Faker==37.11.0
Flask==3.1.2
gender-guesser==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
pandas==2.3.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
